### 2024-06-14 #6890 @editorialbot
[PRE REVIEW]: lintsampler: Easy random sampling via linear interpolation
**Submitting author:** @aneeshnaik (<a href="http://orcid.org/0000-0001-6841-1496">Aneesh Naik</a>)
**Repository:** https://github.com/aneeshnaik/lintsampler
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
