### 2024-06-14 #6890 @jbytecode
@aneeshnaik - Hi, I am the handling editor of this submission. Thank you for sending you work to JOSS.

Do you have any suggestions for the potential reviewers? The manuscript and the code is on mainly Statistics and Python so you can search people that have experiences on these subjects using the search tool:

https://reviewers.joss.theoj.org/lookup

When you suggest potential reviewers, please do not use the `@` character when you mention their GitHub handles just because we don't want to bother people.

Thank you in advance.
