### 2024-06-16 #6890 @aneeshnaik
Hi @jbytecode, thanks for agreeing to serve as the editor!

Having searched the reviewer list you sent, here are some initial suggestions for suitable referees. I don't know any of them personally.
- matt-graham
- DanWaxman
- Edenhofer
- vankesteren
- larryshamalama

However, it seemed like there might well be many other suitable choices. Perhaps my co-author @michael-petersen might also have some suggestions?
