### 2024-06-16 #6890 @jbytecode
👋👋👋 Dear @matt-graham and @vankesteren 👋👋👋

[[PRE REVIEW]: lintsampler: Easy random sampling via linear interpolation](https://github.com/openjournals/joss-reviews/issues/6890#top)

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/6890)).

Thank you in advance!
