### 2024-06-19 #6890 @vankesteren
Small question about the review process: was the scope/statement of need already checked by the editorial team before moving to the pre-review stage?

I have reservations about the statement of need, which would be elaborated in my review. If that is ok, I am happy to review!
