### 2024-06-20 #6890 @jbytecode
@vankesteren - Scope check is always considered by our track editors. You can provide your thoughts and suggest corrections during the review process.
