### 2024-06-20 #6906 @editorialbot
[REVIEW]: lintsampler: Easy random sampling via linear interpolation
**Submitting author:** @aneeshnaik (<a href="http://orcid.org/0000-0001-6841-1496">Aneesh Naik</a>)
**Repository:** https://github.com/aneeshnaik/lintsampler
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
