### 2024-06-26 aneeshnaik/lintsampler#6 @vankesteren
JOSS review: add community guidelines
I could not find community guidelines anywhere, and JOSS requires them.

See https://github.com/openjournals/joss-reviews/issues/6906
