### 2024-06-26 aneeshnaik/lintsampler#8 @vankesteren
JOSS review: main theoretical issue
Dear authors,

Hereby an issue to track the main point raised in my review of the work you have submitted to JOSS. See the review thread [here](https://github.com/openjournals/joss-reviews/issues/6906#issuecomment-2181497075).

First of all, let me congratulate you on the excellent technical state of your package. The documentation, tests, code quality are all top-notch. However, in my opinion an important issue needs to be addressed before the work can be considered for publication in JOSS (note that the editor may disagree!)

The main issue I have is with the statement of need. It is unclear, both in your paper and in your documentation, when I would want / need to use `lintsampler`.

First, a better description of relevant real-world research situations is needed. It needs to be better than

> For a small number of well-studied probability distributions, optimised algorithms exist to draw samples cheaply. However, one often wishes to draw samples from an arbitrary PDF for which no such algorithm is available.

As an external reader, I may disagree with the "a small number of well-studied probability distributions", as there is a large number of probability distributions, which cover a wide range of real-world situations from which sampling is relatively easy to do (e.g., your GMM example which is easy to sample from without `lintsampler`). What research problem that you envision does not fall within this set of distributions?

Second, and more pertinent: you do clearly contrast your method to more complex methods such as MCMC, but you fail to compare it to simpler methods like importance sampling. Currently, information is lacking: when is your method preferred over simpler methods? Does it scale well in higher dimensions? Is it computationally more efficient? What is the advantage?

```py
# Example of importance sampling
propdist  = uniform(-12, 24) # or normal(0, 4) or student(2) or ...
proposals = propdist.rvs(1000)
weights   = pdf(proposals) # / propdist.pdf(proposals) not needed because uniform!
samples   = np.random.choice(proposals, 1000, p=weights / weights.sum())
```

In particular, I was curious about whether importance sampling could even be better than `lintsampler` for the same number of samples. The results from my tests — using the examples from your documentation — indeed seem to indicate that importance sampling with a uniform proposal distribution can perform better in those cases. The advantage will only grow with better proposal distributions. The comparison is available in [this GitHub gist](https://gist.github.com/vankesteren/1854cf92b6ec26a3bb9628048bb2b9b6).
