### 2024-06-27 aneeshnaik/lintsampler#10 @matt-graham
Minor issues in JOSS paper text
Raising as part of JOSS review openjournals/joss-reviews/issues/6906

Some small typos and other minor suggestions with regards to the JOSS paper text

- [x] Probability density function (PDF) acronym should be defined on first usage.
- [x] _PDF function_ is used in several places but would expand to _probability density function function_.
- [x] Citations should be added for both [NumPy ](https://numpy.org/citing-numpy/) and [Scipy ](https://scipy.org/citing-scipy/).
- [x] In _Statement of need_, _Markov Chain Monte Carlo_ should ideallly be capitalized as _Markov chain Monte Carlo_ (generally the full form of an acronym or initialism will be capitalized as it would usually be, and while _Markov_ and _Monte Carlo_ are both proper nouns, _chain_ is not).
- [x] The SGMCMCJax bibtex entry needs additional braces adding so  _markov chain monte carlo_ is capitalized as _Markov chain Monte Carlo_.
