### 2024-06-27 aneeshnaik/lintsampler#8 @matt-graham
I agree with @vankesteren's comments here and would also like to see the statement of need clarified - I would also echo @vankesteren point that the technical quality of the package here is great, and it might be that the points were are raising here are out of scope for the JOSS review process.

In

> We anticipate `lintsampler` finding use in many applications in scientific research and other areas underpinned by statistics. In such fields, pseudo-random sampling fulfils a myriad of purposes, such as Monte Carlo integration, Bayesian inference, or the generation of initial conditions for numerical simulations. The linear interpolant sampling algorithm underpinning `lintsampler` is a simple and effective alternative to existing techniques, and has no publicly available implementation at present.

you state that pseudo-random sampling can be used for purposes such as Monte Carlo integration or Bayesian inference. To my mind the latter is a specialised subset of the former - the end goal of sampling in Bayesian inference is typically to form Monte Carlo estimates of intractable integrals such as posterior expectations or model evidence terms. In the context of estimating integrals, the linear interpolant sampling algorithm implemented in `lintsampler`, as it relies on some form of tensor product grid on the space being integrated over, I believe will have a computational cost (in terms of the number of evaluations of the integrand / density function)  which scales exponentially in the dimension of the space to achieve a given approximation error. This I think puts linear interpolant sampling in the same computational complexity bracket as numerical quadrature rules for multidimensional integrals based on tensor product grids. Given there are a wide variety of numerical quadrature rules available with well understood numerical convergence properties and a range of Python implementations available, my question would therefore be when / in what sort of problems should a user approximate an integral using linear interpolant sampling rather than a numerical quadrature method?
