### 2024-07-01 #6906 @aneeshnaik
@jbytecode: Thank you!

@vankesteren & @matt-graham: Thank you both so much for your work reviewing our code and paper. Your comments all seem quite reasonable and well-judged, and they will serve to improve `lintsampler` and the paper. Here is our plan to address your comments:

- For the 'minor' issues (6, 7, 9, 10), I have created a single [PR](https://github.com/aneeshnaik/lintsampler/pull/11) that will resolve them all. I'll ping you again once we have ticked all the checkboxes and the PR is on the point of merging.
- For the 'major' issue (8) regarding the statement of need, comparison to importance sampling etc., we're working on a slightly more quantitative response, possibly a notebook, that will more concretely identify where/when/why a user might want to use `lintsampler` over alternatives. We'll run that by you, then update the statement of need and the docs subsequently.
