+++ README.md
@@ -34,6 +34,7 @@ If you have a density function, such as this multi-modal 1d pdf with the bulk of
 
 ```python
 import numpy as np
+from scipy.stats import norm
 
 def gmm_pdf(x):
     mu = np.array([-3.0, 0.5, 2.5])
