### 2024-07-05 eaa29bd
Docs: add 'when to use' page
