### 2024-07-07 f75da2d
Add gaussian example to JOSS paper
