### 2024-07-08 170613d
Docs: add real world example to 'when to use' section
