+++ docs/source/whentouse.md
@@ -1,29 +1,49 @@
 # When to Use `lintsampler`
 
-Below is a (non-exhaustive) list of situations where a user might find `lintsampler` (and/or the the linear interpolant sampling algorithm underpinning it) to be preferable over other random sampling techniques.
+Below is a (non-exhaustive) list of 'use cases', i.e., situations where a user might find `lintsampler` (and/or the the linear interpolant sampling algorithm underpinning it) to be preferable over other random sampling techniques. This page closely mirrors the 'Statement of Need' section of our [JOSS paper](https://joss.theoj.org/papers/133f9681a79ab47b916ecee9e183be9e).
 
 In all of these use cases, it is assumed that the dimension of the problem is not too high. `lintsampler` works by evaluating a given PDF on the nodes of a grid (or grid-like structure, such as a tree), so the number of evaluations (and memory occupancy) grows exponentially with the number of dimensions. As a consequence, many of the efficiency arguments given for `lintsampler` below don't apply to higher dimensional problems. We probably wouldn't use `lintsampler` in more than 6 dimensions, but there is no hard limit here: the question of how many dimensions is too many will depend on the problem at hand. 
 
-A second assumption made below is that the target PDF the user wishes to sample from does not have its own exact sampling algorithm (such as the [Box-Muller transform](https://en.wikipedia.org/wiki/Box-Muller_transform) for a Gaussian PDF). If it does, use that instead.
+A second assumption made below is that the target PDF the user wishes to sample from does not have its own exact sampling algorithm (such as the [Box-Muller transform](https://en.wikipedia.org/wiki/Box-Muller_transform) for a Gaussian PDF). The power of `lintsampler` lies in its applicability to arbitrary PDFs for which tailor-made sampling algorithms are not available.
 
 
+## Use Cases
 
-1. Expensive PDF
+### 1. Expensive PDF
    
-   If the PDF being sampling from is expensive to evaluate and large number of samples is desired, then `lintsampler` might be the most cost-effective option. This is because `lintsampler` does not evaluate the PDF for each sample (as would be the case for importance sampling, rejection sampling or MCMC), but on the nodes of the user-chosen grid. Particularly in a low-dimensional setting where the grid does not have too many nodes, this can mean far fewer PDF evaluations. This point is demonstrated in the [first example notebook](./example_notebooks/1_gmm.ipynb).
+If the PDF being sampled from is expensive to evaluate and a large number of samples is desired, then `lintsampler` might be the most cost-effective option. This is because `lintsampler` does not evaluate the PDF for each sample (as would be the case for importance sampling, rejection sampling or MCMC), but on the nodes of the user-chosen grid. Particularly in a low-dimensional setting where the grid does not have too many nodes, this can mean far fewer PDF evaluations. This point is demonstrated in the [first example notebook](./example_notebooks/1_gmm.ipynb).
 
-   Similarly, there might be situations where the user is not so concerned about strict statistical representativeness but wants to generate a huge number of samples from a target PDF with the least possible computational cost (such as e.g., generating realistic point cloud distributions in video game graphics), they can use `lintsampler` with a coarse grid (so minimal PDF evaluations), then `sample()` to their heart's content.
+Similarly, there might be situations where the user is not so concerned about strict statistical representativeness but wants to generate a huge number of samples from a target PDF with the least possible computational cost (such as e.g., generating realistic point cloud distributions in video game graphics). They can use `lintsampler` with a coarse grid (so minimal PDF evaluations), then `sample()` to their heart's content.
 
-2. Multimodal PDF
 
[…]
