### 2024-07-08 ed5136f
Paper: rewrite statement of need (matches whentouse page in docs)
