### 2024-07-08 d4cbc05
Paper: remove extraneous links in statement of need
