+++ paper/paper.bib
@@ -116,3 +116,41 @@ archivePrefix = {arXiv},
  publisher     = {Springer Science and Business Media {LLC}},
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
+
+
+@ARTICLE{agama.paper,
+       author = {{Vasiliev}, Eugene},
+        title = "{AGAMA: action-based galaxy modelling architecture}",
+      journal = {\mnras},
+     keywords = {methods: numerical, galaxies: kinematics and dynamics, Astrophysics - Astrophysics of Galaxies},
+         year = 2019,
+        month = jan,
+       volume = {482},
+       number = {2},
+        pages = {1525-1544},
+          doi = {10.1093/mnras/sty2672},
+archivePrefix = {arXiv},
+       eprint = {1802.08239},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019MNRAS.482.1525V},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{exp.paper,
+       author = {{Petersen}, Michael S. and {Weinberg}, Martin D. and {Katz}, Neal},
+        title = "{EXP: N-body integration using basis function expansions}",
+      journal = {\mnras},
+     keywords = {methods: numerical, Galaxy: halo, galaxies: haloes, galaxies: kinematics and dynamics, galaxies: structure, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2022,
+        month = mar,
+       volume = {510},
+       number = {4},
+        pages = {6201-6217},
+          doi = {10.1093/mnras/stab3639},
+archivePrefix = {arXiv},
+       eprint = {2104.14577},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022MNRAS.510.6201P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
