### 2024-07-09 59f591c
Tidy up paper: add NFW ref, fix broken bullet list
