+++ paper/paper.bib
@@ -7,6 +7,24 @@
   Address                  = {New York, NY, USA}
 }
 
+@ARTICLE{nfw.paper,
+       author = {{Navarro}, Julio F. and {Frenk}, Carlos S. and {White}, Simon D.~M.},
+        title = "{A Universal Density Profile from Hierarchical Clustering}",
+      journal = {\apj},
+     keywords = {Cosmology: Theory, Cosmology: Dark Matter, Galaxies: Halos, Methods: Numerical, Astrophysics},
+         year = 1997,
+        month = dec,
+       volume = {490},
+       number = {2},
+        pages = {493-508},
+          doi = {10.1086/304888},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/9611107},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1997ApJ...490..493N},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{emcee.paper,
        author = {{Foreman-Mackey}, Daniel and {Farr}, Will and {Sinha}, Manodeep and {Archibald}, Anne and {Hogg}, David and {Sanders}, Jeremy and {Zuntz}, Joe and {Williams}, Peter and {Nelson}, Andrew and {de Val-Borro}, Miguel and {Erhardt}, Tobias and {Pashchenko}, Ilya and {Pla}, Oriol},
         title = "{emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC}",
