### 2024-07-09 ca02355
Paper: Update Devroye and pyMC refs
