+++ paper/paper.bib
@@ -4,7 +4,8 @@
   Author                   = {Luc Devroye},
   Publisher                = {Springer-Verlag},
   Year                     = {1986},
-  Address                  = {New York, NY, USA}
+  Address                  = {New York, NY, USA},
+  doi                      = {10.1007/978-1-4613-8643-8}
 }
 
 @ARTICLE{nfw.paper,
@@ -44,15 +45,17 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@software{pymc.paper,
-       author = {{Fonnesbeck}, Chris and {Patil}, Anand and {Huard}, David and {Salvatier}, John},
-        title = "{PyMC: Bayesian Stochastic Modelling in Python}",
- howpublished = {Astrophysics Source Code Library, record ascl:1506.005},
-         year = 2015,
-        month = jun,
-          eid = {ascl:1506.005},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2015ascl.soft06005F},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+@article{pymc.paper,
+ title={PyMC: Bayesian Stochastic Modelling in Python},
+ volume={35},
+ url={https://www.jstatsoft.org/index.php/jss/article/view/v035i04},
+ doi={10.18637/jss.v035.i04},
+ abstract={This user guide describes a &amp;lt;b&amp;gt;Python&amp;lt;/b&amp;gt; package, &amp;lt;b&amp;gt;PyMC&amp;lt;/b&amp;gt;, that allows users to efficiently code a probabilistic model and draw samples from its posterior distribution using Markov chain Monte Carlo techniques.},
+ number={4},
+ journal={Journal of Statistical Software},
+ author={Patil, Anand and Huard, David and Fonnesbeck, Christopher J.},
+ year={2010},
+ pages={1–81}
 }
 
 
