### 2024-07-09 e130ce8
Docs: update 'when to use' section to match statement of need
