### 2024-07-09 aneeshnaik/lintsampler#6 @aneeshnaik
We have added a 'Contributing' section to the README in the `joss-major` branch (PR #12)
