### 2024-07-09 aneeshnaik/lintsampler#10 @aneeshnaik
These issues are all addressed in the joss-major branch (PR #12)
