### 2024-07-09 aneeshnaik/lintsampler#8 @aneeshnaik
In response to these comments, we spent some time coming up with a more quantitative treatment of when exactly a user might wish to use `lintsampler` over an alternative technique. To summarise, we came up with 4 'use cases'. The use cases are distinct, but we nonetheless described a 'real world' scenario (drawn from our own research) that encapsulates all of them at once. We've listed these in the paper's Statement of Need, and also added a page to the docs ('When to use `lintsampler`) which contains largely the same content.

The four use cases are reflected in the four example notebooks in the docs; we've made this explicit in the paper. We've also updated these notebooks to include some code comparisons with alternative techniques. In particular, notebook 1 in the docs now includes a comparison with importance sampling, and notebook 2 now includes comparisons to both importance sampling and MCMC.

You can see all of these changes in the `joss-major` branch of the repo. There is an associated PR (#12) which can be merged into main once you have had a chance to review the changes. To see a built version of the edited docs, you can choose `joss-major` as the 'version' on the readthedocs page (using the interface at the bottom-left corner).

Some more detailed comments responding to some of your points above:
- @vankesteren: in the example you wrote up in your gist, importance sampling outperforms `lintsampler`, but this is because `lintsampler` is being used with a very coarse grid. If you increase the grid resolution, `lintsampler` manages to achieve a comparable log-likelihood to importance sampling. This is actually a point in `lintsampler`'s favour, because you can get there without *too* fine a resolution, and so you can achieve comparable performance to importance sampling with far fewer PDF evaluations. We've written this point up as the first 'use case', and shown this explicitly in the first example notebook.
- @matt-graham: We agree with your assessment that the computational cost scales exponentially with the number of dimensions. We've added a 'caveat' in the Statement of Need indicating that `lintsampler` would rarely be the method of choice for a high-dimensional problem.
