### 2024-07-09 #6906 @aneeshnaik
Update:

We have created a single [PR](https://github.com/aneeshnaik/lintsampler/pull/12), which addresses all 5 of the issues. I've added a comment on each issue thread which goes into a little more detail about what we've done.

The PR corresponds to the `joss-major` branch of the repo --- you can view all of the changes there. You can also see a built version of the edited docs by choosing `joss-major` as the 'version' on the readthedocs page (using the interface at the bottom-left corner).
