### 2024-07-22 #6906 @matt-graham
I've added some comments to @aneeshnaik's PR. The PR addresses all of the minor comments I had about the paper in aneeshnaik/lintsampler#10 (thanks @aneeshnaik for fixing these!) but I still have some follow on questions regarding the points raised by myself and @vankesteren in aneeshnaik/lintsampler#8.
