### 2024-08-05 #6906 @michael-petersen
Hello @jbytecode,

We've appreciated all the thoughtful comments by @matt-graham and @vankesteren; @aneeshnaik and I have responded to comments (https://github.com/aneeshnaik/lintsampler/pull/12#discussion_r1688206813 and https://github.com/aneeshnaik/lintsampler/pull/12#discussion_r1688187324), and are happy to address any further comments. Once those are resolved we are prepared to merge https://github.com/aneeshnaik/lintsampler/pull/12, which will resolve the remaining open issues.
