### 2024-08-26 6f19faf
Add error scaling page to theory section
