### 2024-09-06 #6906 @matt-graham
Hi - apologies for the silence on my side, have been away on leave.  @aneeshnaik and @michael-petersen have done an excellent job at responding to and addressing my comments and suggestions in aneeshnaik/lintsampler#12. I have now checked all the items in my reviewer checklist as from my perspective I am happy to recommend acceptance once aneeshnaik/lintsampler#12 is merged in.
