+++ paper/paper.bib
@@ -9,7 +9,7 @@
 
 @ARTICLE{emcee.paper,
        author = {{Foreman-Mackey}, Daniel and {Farr}, Will and {Sinha}, Manodeep and {Archibald}, Anne and {Hogg}, David and {Sanders}, Jeremy and {Zuntz}, Joe and {Williams}, Peter and {Nelson}, Andrew and {de Val-Borro}, Miguel and {Erhardt}, Tobias and {Pashchenko}, Ilya and {Pla}, Oriol},
-        title = "{emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC}",
+        title = "{emcee v3: A {Python} ensemble sampling toolkit for affine-invariant {MCMC}}",
       journal = {The Journal of Open Source Software},
      keywords = {Python, astronomy, Astrophysics - Instrumentation and Methods for Astrophysics, Statistics - Computation},
          year = 2019,
@@ -28,7 +28,7 @@ archivePrefix = {arXiv},
 
 @software{pymc.paper,
        author = {{Fonnesbeck}, Chris and {Patil}, Anand and {Huard}, David and {Salvatier}, John},
-        title = "{PyMC: Bayesian Stochastic Modelling in Python}",
+        title = "{PyMC: Bayesian Stochastic Modelling in {P}ython}",
  howpublished = {Astrophysics Source Code Library, record ascl:1506.005},
          year = 2015,
         month = jun,
@@ -43,7 +43,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{pxmcmc.paper,
        author = {{Marignier}, Augustin},
-        title = "{PxMCMC: A Python package for proximal Markov Chain Monte Carlo}",
+        title = "{PxMCMC: A {P}ython package for proximal {M}arkov {C}hain {M}onte {C}arlo}",
       journal = {The Journal of Open Source Software},
      keywords = {astrophysics, geophysics, Python, MCMC, imaging},
          year = 2023,
@@ -66,6 +66,6 @@ archivePrefix = {arXiv},
   number = {72},
   pages = {4113},
   author = {Jeremie Coullon and Christopher Nemeth},
-  title = {SGMCMCJax: a lightweight JAX library for stochastic gradient Markov chain Monte Carlo algorithms},
+  title = {SGMCMCJax: a lightweight {JAX} library for stochastic gradient {M}arkov {C}hain {M}onte {C}arlo algorithms},
   journal = {Journal of Open Source Software}
 }
