### 2024-09-13 #6906 @jbytecode
@aneeshnaik - I have just sent a pull request. This PR includes some minor changes in both the manuscript and the bibliography. Please review the changes. If you are agreed with them, please merge. Thank you in advance.

https://github.com/aneeshnaik/lintsampler/pull/13
