### 2024-09-16 323972c
Adjustments to paper for PR#13
