+++ paper/paper.bib
@@ -28,7 +28,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{emcee.paper,
        author = {{Foreman-Mackey}, Daniel and {Farr}, Will and {Sinha}, Manodeep and {Archibald}, Anne and {Hogg}, David and {Sanders}, Jeremy and {Zuntz}, Joe and {Williams}, Peter and {Nelson}, Andrew and {de Val-Borro}, Miguel and {Erhardt}, Tobias and {Pashchenko}, Ilya and {Pla}, Oriol},
-        title = "{emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC}",
+        title = "{emcee v3: A {P}ython ensemble sampling toolkit for affine-invariant {MCMC}}",
       journal = {The Journal of Open Source Software},
      keywords = {Python, astronomy, Astrophysics - Instrumentation and Methods for Astrophysics, Statistics - Computation},
          year = 2019,
@@ -46,7 +46,7 @@ archivePrefix = {arXiv},
 }
 
 @article{pymc.paper,
- title={PyMC: Bayesian Stochastic Modelling in Python},
+ title="{PyMC: Bayesian Stochastic Modelling in {P}ython}",
  volume={35},
  url={https://www.jstatsoft.org/index.php/jss/article/view/v035i04},
  doi={10.18637/jss.v035.i04},
@@ -64,7 +64,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{pxmcmc.paper,
        author = {{Marignier}, Augustin},
-        title = "{PxMCMC: A Python package for proximal {M}arkov chain {M}onte {C}arlo}",
+        title = "{PxMCMC: A {P}ython package for proximal {M}arkov {C}hain {M}onte {C}arlo}",
       journal = {The Journal of Open Source Software},
      keywords = {astrophysics, geophysics, Python, MCMC, imaging},
          year = 2023,
@@ -87,7 +87,7 @@ archivePrefix = {arXiv},
   number = {72},
   pages = {4113},
   author = {Jeremie Coullon and Christopher Nemeth},
-  title = {SGMCMCJax: a lightweight JAX library for stochastic gradient {M}arkov chain {M}onte {C}arlo algorithms},
+  title = {SGMCMCJax: a lightweight {JAX} library for stochastic gradient {M}arkov {C}hain {M}onte {C}arlo algorithms},
   journal = {Journal of Open Source Software}
 }
 
