### 2024-09-20 #6906 @aneeshnaik
Thanks @jbytecode. Your PR was on the old version of lintsampler rather than the updated one (we had not yet merged the changes when you added your edits to the main branch), but we incorporated as many of your edits as possible in the updated version. This was in [PR 14](https://github.com/aneeshnaik/lintsampler/pull/14), which we have now merged into main. I don't know if you might want to take a look at the new version of the paper, as there might be some content in there which you have not yet reviewed.
