### 2024-09-22 #6906 @jbytecode
@aneeshnaik

Please review these changes on the manuscript:

- The line 59, does $N^{-1/2}$ indicate an order of magnitude? Is it better to use an *O* notation there?
- The line 107, **with a interface** should be **with an interface**

After editing the paper please perform the actions given below:
