### 2024-09-24 1b34773
Minor edits to paper.md
