### 2024-09-24 cc78a12
Add JOSS badge and links to README
