### 2024-09-24 4f8c20c
Add JOSS links to docs homepage
