### 2024-09-27 #6906 @aneeshnaik
Hi @jbytecode:

- Implemented your suggested edits in the manuscript.
- Created a tagged release. The version is v1.0.0
- Archived the release on Zenodo (with matching license, authors, ORCIDs). The DOI is 10.5281/zenodo.13846862
